10477 Move NpsIabsSummary reads into companion obj
